refactor: extract MobileRoutinesTab component
